backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
PyYAML==6.0.2
pytest==8.2.2
pytest-django==4.8.0
black==24.4.2
isort==5.13.2